Split generic git-wrangling functions to separate module
